Move v26 blockchain to subdirectory
In preparation for adding new blockchain RPCs move the module to a
subdirectory and split out `error` and `into` into modules.

Code move only, no functional changes.
